src/api_service/requirements.txt
fastapi
uvicorn
pyjwt
passlib[bcrypt]
grpcio
grpcio-tools
protobuf
